server/requirements.txt
# EarthRing Procedural Generation Service
# Python dependencies for the procedural generation service

# Core dependencies
numpy>=1.26.0
scipy>=1.12.0

# Database
psycopg2-binary>=2.9.9

# API framework
FastAPI>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
httpx>=0.27.0
black>=24.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0

